splits off search into it's own file
